+++ CHANGELOG.md
@@ -17,7 +17,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - is_time_lag: Checks to see if there is a time lag.
 
 ### Changed
-- chk_slv_err: Added "Status" key to IO and SQL error dictionaries.
+- chk_slv_err, \_chk_other: Added "Status" key to IO and SQL error dictionaries.
 - call_run_chk: Added check to see if -x option is set and if there is a time lag detected.
 - data_out: Added option to send_mail call to override postfix and use mailx command.
 - config/slave.txt.TEMPLATE: Added TLS entry.